Makefile: Fix lwip library build when building STANDALONE=y.
When building STANDALONE=y toolchain fails to build do to lwip library
failing to build. Lwip library file "liblwip_open.a" fails to build
because its looking for xtensa-lx106-elf-ar in wrong path. Proposed fix
is to pass the correct AR path to esp-open-lwip/Makefile.open as it was
done with CC path.